+++ test/mini_magick_test.rb
@@ -2,19 +2,23 @@ require "test_helper"
 require "image_processing/mini_magick"
 require "pathname"
 
-MiniMagick.cli = :graphicsmagick if ENV["GM"]
+if ENV["GM"]
+  MiniMagick.configure do |config|
+    config.cli_prefix = "gm"
+  end
+end
 
 describe "ImageProcessing::MiniMagick" do
   before do
     @portrait  = fixture_image("portrait.jpg")
     @landscape = fixture_image("landscape.jpg")
-    @square = fixture_image("square.jpg")
+    @square    = fixture_image("square.jpg")
   end
 
   it "applies imagemagick operations" do
     actual = ImageProcessing::MiniMagick.flip.call(@portrait)
     expected = Tempfile.new(["result", ".jpg"], binmode: true).tap do |tempfile|
-      MiniMagick::Tool::Convert.new do |cmd|
+      MiniMagick.convert do |cmd|
         cmd << @portrait.path
         cmd.flip
         cmd << tempfile.path
@@ -37,7 +41,7 @@ describe "ImageProcessing::MiniMagick" do
   it "applies macro operations" do
     actual = ImageProcessing::MiniMagick.resize_to_limit(400, 400).call(@portrait)
     expected = Tempfile.new(["result", ".jpg"], binmode: true).tap do |tempfile|
-      MiniMagick::Tool::Convert.new do |cmd|
+      MiniMagick.convert do |cmd|
         cmd << @portrait.path
         cmd.resize("400x400")
         cmd << tempfile.path
@@ -55,7 +59,7 @@ describe "ImageProcessing::MiniMagick" do
 
   it "accepts page" do
     tiff = Tempfile.new(["file", ".tiff"])
-    MiniMagick::Tool::Convert.new do |convert|
+    MiniMagick.convert do |convert|
       convert.merge! [@portrait.path, @portrait.path, @portrait.path]
       convert << tiff.path
     end
@@ -70,7 +74,7 @@ describe "ImageProcessing::MiniMagick" do
 
   it "disallows split layers by default" do
     tiff = Tempfile.new(["file", ".tiff"])
-    MiniMagick::Tool::Convert.new do |convert|
+    MiniMagick.convert do |convert|
       convert.merge! [@portrait.path, @portrait.path, @portrait.path]
       convert << tiff.path
     end
@@ -168,7 +172,7 @@ describe "ImageProcessing::MiniMagick" do
   end
 
   it "accepts magick object as source" do
-    magick = MiniMagick::Tool::Convert.new
+    magick = MiniMagick.convert
     magick << fixture_image("rotated.jpg").path
     result = ImageProcessing::MiniMagick.source(magick).call
     assert_dimensions [600, 800], result
@@ -588,7 +592,7 @@ describe "ImageProcessing::MiniMagick" do
     it "appends CLI arguments" do
       actual = ImageProcessing::MiniMagick.append("-resize", "400x400").call(@portrait)
       expected = Tempfile.new(["result", ".jpg"], binmode: true).tap do |tempfile|
-        MiniMagick::Tool::Convert.new do |cmd|
+        MiniMagick.convert do |cmd|
           cmd << @portrait.path
           cmd.resize("400x400")
           cmd << tempfile.path